feat(graphqlArg): allow specify explicit type
